new lib

+++ Libft/ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: hdecaux <hdecaux@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2015/10/09 12:47:37 by hdecaux           #+#    #+#             */
-/*   Updated: 2015/10/09 12:48:19 by hdecaux          ###   ########.fr       */
+/*   Updated: 2016/11/12 17:11:21 by hdecaux          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *s)
 {
-	int	a;
+	size_t	a;
 
 	a = 0;
 	while (s[a])
